bogdan-project/requirements.txt
Django==4.0.6
djangorestframework==3.13.1
Pillow==9.2.0
psycopg2-binary